ESQUIZOFRENIA TOTAAAAAAAAAAAAAAAAAAAAAAAAL

+++ shared-2/src/shared-2.c
@@ -55,7 +55,7 @@ int esperar_cliente_alt(int socket_servidor) {
 int recibir_operacion(int socket_cliente)
 {
 	int cod_op;
-	if(recv(socket_cliente, &cod_op, sizeof(uint8_t), MSG_WAITALL) > 0)
+	if(recv(socket_cliente, &cod_op, sizeof(int), MSG_WAITALL) > 0)
 		return cod_op;
 	else
 	{
